pro: Added support for multi-level provisioning
For RocksDB, SSTables are stored in zones according to the LSM-Tree
level. In this way, a single zone contains only SSTables belonging
to the same level. When compaction takes place, we invalidade entire
zones and avoid moving valid data, decreasing write-amplification.

+++ xapp/include/xapp.h
@@ -95,7 +95,7 @@ struct xapp_io_ucmd {
     uint64_t 	   id;
     void	  *buf;
     uint32_t 	   size;
-    uint8_t 	   prov_type;
+    uint16_t 	   prov_type;
     uint8_t 	   app_md;  /* Application is responsible for mapping/recovery */
     uint8_t 	   status;
